// src/api/gen/types/CustomerRequest.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type CustomerRequest = {
    /**
     * @description
     * Format: `date`
     * @type string | undefined
    */
    birthDate?: string;
    email?: string;
    /**
     * @minLength 1
     * @type string
    */
    name: string;
    nationalId?: string;
    /**
     * @minLength 1
     * @type string
    */
    phone: string;
    postalCode?: string;
    street?: string;
};